// src/types/position.ts
/*
 * @Author: Henry Ma henryma@edening.cn
 * @Date: 2025-01-21 15:45:08
 * @LastEditors: Henry Ma henryma@edening.cn
 * @LastEditTime: 2025-01-21 15:45:28
 * @FilePath: /GlobalVillage/src/types/position.ts
 * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
 */
export interface Position {
  LNG: number
  LAT: number
  ALT: number
}